--- go/wallet/cmd/api_wasm.go
@@ -1,20 +1,0 @@
-//go:build js && wasm
-// +build js,wasm
-
-package main
-
-import (
-	"fmt"
-
-	"github.com/onflow/fcl-dev-wallet/go/wallet"
-)
-
-func main() {
-	srv, err := wallet.NewApiHTTPServer(nil)
-	if err != nil {
-		panic(err)
-	}
-
-	fmt.Println("Development server started on port 8701")
-	srv.Start()
-}
